Include fallback functions only if using Lumen. Fix #1377
